requirements.txt
pytest==7.4.0
pytest-html==3.2.0
pytest-metadata==3.0.0
allure-pytest==2.13.2
PyYAML==6.0.1
selenium==3.14.1
zmail==0.2.8
appium-python-client==1.2.0